refactor: move coaches to app
