+++ apps/front/src/app/coaches/page.tsx
@@ -1,30 +1,29 @@
-"use client";
+'use client';
 import React, { useState } from 'react';
 import { Page } from '@seminar/common';
-import CoachesTableHeader from './table/header';
 import CoachesTable from './table';
 import { Coach } from '@/types/coach';
 
 export default function Coaches() {
-    const [coachesPage, setCoachesPage] = useState<Page<Coach>>({
-      index: 0,
-      size: 15,
-      isLast: false,
-      items: temporaryCoaches,
-    });
-    const [numberOfCoaches, setNumberOfCoaches] = useState<number>(temporaryCoaches.length);
+  const [coachesPage, setCoachesPage] = useState<Page<Coach>>({
+    index: 0,
+    size: 15,
+    isLast: false,
+    items: temporaryCoaches,
+  });
+  const [numberOfCoaches, setNumberOfCoaches] = useState<number>(temporaryCoaches.length);
 
-    return <div>
-      <div className='p-8'>
-        <h1 className="scroll-m-20 text-2xl font-extrabold tracking-tight lg:text-5xl mb-1">
-          Coaches List
-        </h1>
-        <h3 className="text-lg font-bold mb-4">
-          You have total {numberOfCoaches} coaches.
-        </h3>
-        <CoachesTable coaches={coachesPage.items} />
-      </div>
-    </div>;
+  return <div>
+    <div className='p-8'>
+      <h1 className="scroll-m-20 text-2xl font-extrabold tracking-tight lg:text-5xl mb-1">
+        Coaches List
+      </h1>
+      <h3 className="text-lg font-bold mb-4">
+        You have total {numberOfCoaches} coaches.
+      </h3>
+      <CoachesTable coaches={coachesPage.items} />
+    </div>
+  </div>;
 }
 
 const temporaryCoaches: Coach[] = [
