+++ apps/front/src/app/page.tsx
@@ -1,9 +1,13 @@
-import Coaches from "@/components/coaches";
+import { Button } from "@/components/ui/button";
+import Link from "next/link";
 
 export default function Home() {
+
   return (
     <main>
-      <Coaches />
+     <Link href="/coaches">
+        <Button>Coaches</Button>
+      </Link>
     </main>
   );
 }
