+++ apps/front/src/components/ui/avatar.tsx
@@ -1,4 +1,4 @@
-"use client"
+'use client'
 
 import * as React from "react"
 import * as AvatarPrimitive from "@radix-ui/react-avatar"
