+++ apps/front/src/components/ui/checkbox.tsx
@@ -1,4 +1,4 @@
-"use client"
+'use client'
 
 import * as React from "react"
 import * as CheckboxPrimitive from "@radix-ui/react-checkbox"
