+++ apps/front/src/components/ui/dropdown-menu.tsx
@@ -1,4 +1,4 @@
-"use client"
+'use client'
 
 import * as React from "react"
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu"
