+++ apps/front/src/components/ui/label.tsx
@@ -1,4 +1,4 @@
-"use client"
+'use client'
 
 import * as React from "react"
 import * as LabelPrimitive from "@radix-ui/react-label"
@@ -13,7 +13,7 @@ const labelVariants = cva(
 const Label = React.forwardRef<
   React.ElementRef<typeof LabelPrimitive.Root>,
   React.ComponentPropsWithoutRef<typeof LabelPrimitive.Root> &
-    VariantProps<typeof labelVariants>
+  VariantProps<typeof labelVariants>
 >(({ className, ...props }, ref) => (
   <LabelPrimitive.Root
     ref={ref}
