+++ apps/front/src/components/ui/toast.tsx
@@ -1,4 +1,4 @@
-"use client"
+'use client'
 
 import * as React from "react"
 import * as ToastPrimitives from "@radix-ui/react-toast"
@@ -43,7 +43,7 @@ const toastVariants = cva(
 const Toast = React.forwardRef<
   React.ElementRef<typeof ToastPrimitives.Root>,
   React.ComponentPropsWithoutRef<typeof ToastPrimitives.Root> &
-    VariantProps<typeof toastVariants>
+  VariantProps<typeof toastVariants>
 >(({ className, variant, ...props }, ref) => {
   return (
     <ToastPrimitives.Root
