+++ apps/front/src/components/ui/toaster.tsx
@@ -1,4 +1,4 @@
-"use client"
+'use client'
 
 import { useToast } from "@/hooks/use-toast"
 import {
